deviding views: filter_views
